requirements.txt
feedparser==6.0.11
groq>=1.1.2
python-telegram-bot==21.6
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
